We returned the future at the first failed peer exchange attempt which can lead to the situation that the get inventory request is not done because there are no peers yet connected.
Now we complete the future either if all peer exchange attempts have failed or if at least one has succeeded.

+++ network/network/src/main/java/bisq/network/p2p/services/peergroup/exchange/PeerExchangeStrategy.java
@@ -120,7 +120,7 @@ public class PeerExchangeStrategy {
         peerGroup.addReportedPeers(filtered);
     }
 
-    boolean redoInitialPeerExchange(long numSuccess, int numRequests) {
+    boolean redoInitialPeerExchange(int numSuccess, int numRequests) {
         boolean moreThenHalfFailed = numRequests - numSuccess > numRequests / 2;
         return moreThenHalfFailed ||
                 !sufficientConnections() ||
